✨ feat[frontend]: added cookie handler for branding setting and refactored authentication cookie for better naming convention
- Changed name of `cookieSetter` and `cookieDestroy` by adding auth prefix before the names.
- Made a preliminary cookie handler to take care of branding settings of the application
✨ Refactor branding cookie functions

Update function names to better reflect the purpose of setting and getting
the branding color cookie. Ensure consistency and clarity in the codebase.
✨ Update authCookieSetter in GoogleSignIn and GitHubSignIn components

- Update cookieSetter to authCookieSetter for setting authentication cookies in
GoogleSignIn and GitHubSignIn components. This change ensures consistency in
naming conventions and improves readability.
