+++ apps/admin/app/home/(auth)/register/_components/form.tsx
@@ -14,7 +14,7 @@ import {
   TRegistrationErrorResponse,
   TRegistrationSuccessResponse,
 } from "@/types";
-import { cookieSetter } from "@/actions";
+import { authCookieSetter } from "@/actions";
 import { formSchema, TFormSchema } from "./form-schema";
 
 export function RegisterationForm() {
@@ -41,7 +41,7 @@ export function RegisterationForm() {
         REGISTER_URL,
         data
       );
-      cookieSetter({
+      authCookieSetter({
         access: response.data.token.access,
         refresh: response.data.token.refresh,
       });
